pyproject.toml
[project]
name = "ha-payment-gateway"
version = "1.0.0"
description = "Gateway de pagamentos simulado com foco em alta disponibilidade."
readme = "README.html"
license = { text = "MIT" }
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "prometheus-client>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.8.0",
    "coverage>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
ignore = [
    "B008", # Depends() em defaults é o padrão do FastAPI
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "migrations/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
